+++ frontend/src/metabase/query_builder/components/notebook/steps/JoinStep/JoinConditionDraft/utils.ts
@@ -1,0 +1,9 @@
+import * as Lib from "metabase-lib";
+
+export function getDefaultJoinConditionOperator(
+  query: Lib.Query,
+  stageIndex: number,
+): Lib.JoinConditionOperator {
+  const operators = Lib.joinConditionOperators(query, stageIndex);
+  return operators[0];
+}
